@@ -1,6 +1,6 @@
 ﻿using System;
-using System.Net;
-using System.Text;
+using System.Collections.Generic;
+using System.Net.Http;
 using System.Text.Json;
 using System.Text.Json.Serialization;
 using System.Threading.Tasks;
@@ -21,24 +21,18 @@ namespace Loon.Services
 
             try
             {
-                var request = WebRequest.Create(Endpoint);
+                var request = new HttpRequestMessage(HttpMethod.Post, Endpoint);
 
-                var q = "q=" + Uri.EscapeDataString(text);
-                var s = "&source=" + Uri.EscapeDataString(fromLanguage);
-                var t = "&target=" + Uri.EscapeDataString(toLanguage);
-                var bytes = Encoding.UTF8.GetBytes(q + s + t);
+                var content = new FormUrlEncodedContent(new[]
+                {
+                    new KeyValuePair<string?, string?>("q", Uri.EscapeDataString(text)),
+                    new KeyValuePair<string?, string?>("source", Uri.EscapeDataString(fromLanguage)),
+                    new KeyValuePair<string?, string?>("target", Uri.EscapeDataString(toLanguage))
+                });
 
-                request.Method = "POST";
-                request.ContentType = "application/x-www-form-urlencoded";
-                request.ContentLength = bytes.Length;
-
-                var content = await request.GetRequestStreamAsync().ConfigureAwait(false);
-                await content.WriteAsync(bytes).ConfigureAwait(false);
-                content.Close();
-
-                using var response = await request.GetResponseAsync().ConfigureAwait(false);
-                await using var stream = response.GetResponseStream();
-                var result = await JsonSerializer.DeserializeAsync<TranslatorResult>(stream).ConfigureAwait(false);
+                using var response = await Twitter.Services.OAuthApiRequest.MyHttpClient.SendAsync(request).ConfigureAwait(false);
+                var       stream   = await response.Content.ReadAsStreamAsync().ConfigureAwait(false);
+                var       result   = await JsonSerializer.DeserializeAsync<TranslatorResult>(stream).ConfigureAwait(false);
                 return result?.TranslatedText?.HtmlDecode() ?? "{error}";
             }
             catch (Exception ex)
@@ -50,7 +44,6 @@ namespace Loon.Services
 
     public class TranslatorResult
     {
-        [JsonPropertyName("translatedText")]
-        public string? TranslatedText { get; set; }
+        [JsonPropertyName("translatedText")] public string? TranslatedText { get; set; }
     }
 }
