@@ -1,6 +1,5 @@
 ﻿using System;
-using System.IO;
-using System.Text;
+using System.Net.Http;
 using System.Threading.Tasks;
 
 namespace Twitter.Services
@@ -12,22 +11,21 @@ namespace Twitter.Services
             string consumerKey,
             string consumerSecret)
         {
-            const string requestTokenUrl = "https://api.twitter.com/oauth/request_token";
-            var nonce = OAuth.Nonce();
-            var timestamp = OAuth.TimeStamp();
-            var parameters = new[] { ("oauth_callback", "oob") };
-            var signature = OAuth.Signature(OAuthApiRequest.POST, requestTokenUrl, nonce, timestamp, consumerKey, consumerSecret, "", "", parameters);
-            var authorizationHeader = OAuth.AuthorizationHeader(nonce, timestamp, consumerKey, accessToken: null, signature, parameters);
+            const string requestTokenUrl     = "https://api.twitter.com/oauth/request_token";
+            var          nonce               = OAuth.Nonce();
+            var          timestamp           = OAuth.TimeStamp();
+            var          parameters          = new[] { ("oauth_callback", "oob") };
+            var          signature           = OAuth.Signature(OAuthApiRequest.POST, requestTokenUrl, nonce, timestamp, consumerKey, consumerSecret, "", "", parameters);
+            var          authorizationHeader = OAuth.AuthorizationHeader(nonce, timestamp, consumerKey, accessToken: null, signature, parameters);
 
-            var request = System.Net.WebRequest.Create(new Uri(requestTokenUrl));
-            request.Method = OAuthApiRequest.POST;
+            var request = new HttpRequestMessage(HttpMethod.Post, new Uri(requestTokenUrl));
             request.Headers.Add("Authorization", authorizationHeader);
-            using var response = await request.GetResponseAsync().ConfigureAwait(false);
-            using var stream = new StreamReader(response.GetResponseStream(), Encoding.UTF8);
-            var body = await stream.ReadToEndAsync().ConfigureAwait(false);
-            var tokens = body.Split('&');
-            var oauthToken = Token(tokens[0]);
-            var oauthSecret = Token(tokens[1]);
+            using var response = await OAuthApiRequest.MyHttpClient.SendAsync(request).ConfigureAwait(false);
+
+            var body              = await response.Content.ReadAsStringAsync().ConfigureAwait(false);
+            var tokens            = body.Split('&');
+            var oauthToken        = Token(tokens[0]);
+            var oauthSecret       = Token(tokens[1]);
             var callbackConfirmed = Token(tokens[2]);
 
             if (string.CompareOrdinal(callbackConfirmed, "true") != 0)
@@ -37,8 +35,8 @@ namespace Twitter.Services
 
             return new OAuthTokens
             {
-                OAuthToken = oauthToken,
-                OAuthSecret = oauthSecret,
+                OAuthToken  = oauthToken,
+                OAuthSecret = oauthSecret
             };
         }
 
@@ -49,27 +47,26 @@ namespace Twitter.Services
             string accessTokenSecret,
             string oauthVerifier)
         {
-            const string requestTokenUrl = "https://api.twitter.com/oauth/access_token";
-            var nonce = OAuth.Nonce();
-            var timestamp = OAuth.TimeStamp();
-            var parameters = new[] { ("oauth_verifier", oauthVerifier) };
-            var signature = OAuth.Signature(OAuthApiRequest.POST, requestTokenUrl, nonce, timestamp, consumerKey, consumerSecret, accessToken, accessTokenSecret, parameters);
-            var authorizationHeader = OAuth.AuthorizationHeader(nonce, timestamp, consumerKey, accessToken, signature, parameters);
+            const string requestTokenUrl     = "https://api.twitter.com/oauth/access_token";
+            var          nonce               = OAuth.Nonce();
+            var          timestamp           = OAuth.TimeStamp();
+            var          parameters          = new[] { ("oauth_verifier", oauthVerifier) };
+            var          signature           = OAuth.Signature(OAuthApiRequest.POST, requestTokenUrl, nonce, timestamp, consumerKey, consumerSecret, accessToken, accessTokenSecret, parameters);
+            var          authorizationHeader = OAuth.AuthorizationHeader(nonce, timestamp, consumerKey, accessToken, signature, parameters);
 
-            var request = System.Net.WebRequest.Create(new Uri(requestTokenUrl));
-            request.Method = OAuthApiRequest.POST;
+            var request = new HttpRequestMessage(HttpMethod.Post, new Uri(requestTokenUrl));
             request.Headers.Add("Authorization", authorizationHeader);
 
-            using var response = await request.GetResponseAsync().ConfigureAwait(false);
-            using var stream = new StreamReader(response.GetResponseStream(), Encoding.UTF8);
-            var tokens = (await stream.ReadToEndAsync().ConfigureAwait(false)).Split('&');
+            using var response = await OAuthApiRequest.MyHttpClient.SendAsync(request).ConfigureAwait(false);
+            var       content  = await response.Content.ReadAsStringAsync().ConfigureAwait(false);
+            var       tokens   = content.Split('&');
 
             var oauthTokens = new OAuthTokens
             {
-                OAuthToken = Token(tokens[0]),
+                OAuthToken  = Token(tokens[0]),
                 OAuthSecret = Token(tokens[1]),
-                UserId = Token(tokens[2]),
-                ScreenName = Token(tokens[3]),
+                UserId      = Token(tokens[2]),
+                ScreenName  = Token(tokens[3])
             };
 
             return oauthTokens;
